Cleanup and hopefully finalize agent 5
